+++ surfkit.go
@@ -6,7 +6,7 @@ import (
 	"github.com/helloink/surfkit/events"
 )
 
-const version = "1.6.0-x3"
+const version = "1.6.0-x4"
 
 // Output defines the single channel on which the service produces output, given it is a Pubsub output.
 // Eventuall this should also cover HTTP Endpoints.